Moved key exposing to its own decorator
